fix xpu test error

+++ python/paddle/fluid/tests/unittests/xpu/test_pad3d_op_xpu.py
@@ -188,7 +188,9 @@ class XPUTestPad3dOp(XPUOpTestWrapper):
                 mode = "constant"
                 value = 100
                 input_data = np.random.rand(*input_shape).astype(self.dtype)
-                x = paddle.static.data(name="x", shape=input_shape)
+                x = paddle.static.data(
+                    name="x", shape=input_shape, dtype="float32"
+                )
                 result = F.pad(
                     x=x, pad=pad, value=value, mode=mode, data_format="NCDHW"
                 )
@@ -209,7 +211,9 @@ class XPUTestPad3dOp(XPUOpTestWrapper):
                 pad = [1, 2, 1, 1, 1, 2]
                 mode = "reflect"
                 input_data = np.random.rand(*input_shape).astype(self.dtype)
-                x = paddle.static.data(name="x", shape=input_shape)
+                x = paddle.static.data(
+                    name="x", shape=input_shape, dtype="float32"
+                )
                 result1 = F.pad(x=x, pad=pad, mode=mode, data_format="NCDHW")
                 result2 = F.pad(x=x, pad=pad, mode=mode, data_format="NDHWC")
                 exe = Executor(place)
